fix: Move new globals to own file so that they are lazily compiled
Reverted change in classExporter to ignore !public members to only ignoring private ones
