+++ bindgen/assembly/index.ts
@@ -1,5 +1,8 @@
-import { base64, runtime_api, u128, util, Context, storage } from "near-sdk-core";
+import { base64, runtime_api, u128, util, Context } from "near-sdk-core";
 import { JSONEncoder as _JSONEncoder, JSON } from "assemblyscript-json";
+
+// Import globals needed for singleton
+import "./singleton";
 
 // Runtime functions
 // tslint:disable: no-unsafe-any
@@ -465,36 +468,4 @@ function defaultValue<T>(): T {
   }
   return changetype<T>(0);
 }
-// @ts-ignore
-@global
-/* eslint-disable indent */
-const __STATE_KEY = "STATE";
 
-// @ts-ignore
-@global
-function __checkState(): bool {
-  return storage.contains(__STATE_KEY);
-}
-
-// @ts-ignore
-@global
-function __getState<T>(): T {
-  return storage.getSome<T>(__STATE_KEY);
-}
-
-// @ts-ignore
-@global
-let __updated = false;
-// @ts-ignore
-
-@global
-function __setState<T>(state: T): void {
-    storage.set(__STATE_KEY, state);
-}
-
-@global
-function __updateState<T>(state: T): void {
-  if (__updated) {
-    __setState(state);
-  }
-}
